% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateft.R
\name{calculateft}
\alias{calculateft}
\title{Calculate Frequency Tables by Groups}
\usage{
calculateft(frequency_matrix, group_names)
}
\arguments{
\item{frequency_matrix}{A matrix that include frequency data}

\item{group_names}{A name vector that identify each group}
}
\value{
A list with filtered frequency tables per groups
}
\description{
Calculate Frequency Tables by Groups
}
\examples{
frequency_matrix <- matrix(c(1, 2, 2, 3, 3, 1, 4, 1), nrow = 4, byrow = TRUE)
colnames(frequency_matrix) <- c("A", "B")
group_names <- c("Group1", "Group2", "Group3", "Group4")
calculateft (frequency_matrix, group_names)

}
