% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indval1.R
\name{indval1}
\alias{indval1}
\title{Calculate IndVal Index for Taxa in a Group of Samples}
\usage{
indval1(x)
}
\arguments{
\item{x}{x is a matrix with samples in rows and taxa in columns}
}
\value{
index calculated as the product of A and B, multiplied by 100
}
\description{
Auxiliary function for calculating indvalgp for a group of samples
}
\examples{
set.seed(123)
sample_matrix <- matrix(sample(0:5, 30, replace = TRUE), nrow = 6, ncol = 5)
colnames(sample_matrix) <- paste0("Taxon_", 1:5)
rownames(sample_matrix) <- paste0("Sample_", 1:6)

# Mostrar la matriz de ejemplo
print("Matriz de muestras:")
print(sample_matrix)

# Calcular el índice IndVal utilizando la función indval1
indval_result <- indval1(sample_matrix)

# Mostrar el resultado
print("Índice IndVal por taxón:")
print(indval_result)

}
