% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderft.R
\name{orderft}
\alias{orderft}
\title{Order Frequency Tables by Decreasing Values}
\usage{
orderft(frequency_tables)
}
\arguments{
\item{frequency_tables}{A list of data frames, where each data frame represents a frequency table.}
}
\value{
A list of data frames with the rows ordered in decreasing order of the first column values.
}
\description{
This function orders the values of the frequency tables from highest to lowest in each group.
}
\examples{
freq_list <- list(
  data.frame(freq = c(5, 3, 8), row.names = c("A", "B", "C")),
  data.frame(freq = c(2, 6, 1), row.names = c("D", "E", "F"))
)
orderft(freq_list)

}
