% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomchi.R
\name{randomchi}
\alias{randomchi}
\title{Perform Random Chi-Square Tests for Frequency Tables}
\usage{
randomchi(frequency_tables, bioregions, n_iter = 1000)
}
\arguments{
\item{frequency_tables}{A list of frequency tables.}

\item{bioregions}{A dataframe containing a column with any bioregion type used to generate random samples.}

\item{n_iter}{The number of iterations for the random sampling (default is 1000).}
}
\value{
The function returns a list with two elements:
random_table: A list of chi-square test results for each random sample.
p_means: A list of mean p-values for each iteration across all frequency tables.
}
\description{
Perform Random Chi-Square Tests for Frequency Tables
}
\examples{

set.seed(123)

# Example frequency tables (simulated)
 frequency_tables <- list(
 data.frame(A = c(10, 20, 30), B = c(5, 15, 25)),
 data.frame(A = c(10, 20, 30), B = c(10, 20, 30))
)
#Simulating the 'bioregions' dataframe with a 'Province' column
bioregion <- data.frame(Province = sample(letters, 100, replace = TRUE))
# Call the function
result <- randomchi(frequency_tables, bioregion, n_iter = 1000)
# View the results
print(result$p_means)


}
